Create help.py and update en.json (#9)
* Create help.py and update en.json

This is what I can up with for the help command. I did what you said and am drawing the descriptions from the json file. Let me know if I should change something. 

Note: The Wikipedia link that contains the formatting for timezone is also in the json file. I put it there in case if you expand to more languages you can put the articles in the there in a different language.

* Delete en.json

I uploaded the file to the wrong place.

* Update en.json

Because I uploaded the file in the wrong place.

* Fixed error in line 168

* Update help.py

Fixed some issues with embeds missing grave accent, so they weren't code blocks in discord.

* Made sure `user.get_guild()` was calling the correct method.